# workdir/docs/swagger.yaml
basePath: /
definitions:
...
  errors.Code:
    enum:
    - invalid_argument
    - unauthorized
    - forbidden
    - not_found
    - conflict
    - internal
    type: string
    x-enum-varnames:
    - CodeInvalidArgument
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeConflict
    - CodeInternal
  errors.HTTPErrorResponse:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/errors.Code'
        example: not_found
      message:
        example: form not found
        type: string
    type: object
...
  handlers.FormDetailsResponse:
    properties:
      accept_anonymous:
        example: true
        type: boolean
      content:
        type: object
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      description:
        example: Get in touch with our team
        type: string
      end_date:
        example: "2026-12-31T23:59:59Z"
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      public_available:
        example: true
        type: boolean
      slug:
        example: contact-us
        type: string
      start_date:
        example: "2026-01-01T00:00:00Z"
        type: string
      tenant_id:
        example: 1
        type: integer
      tenant_name:
        example: Acme Inc.
        type: string
      title:
        example: Contact us
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    type: object
  handlers.FormListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/handlers.FormSummaryResponse'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 1
        type: integer
    type: object
  handlers.FormRequest:
    properties:
      accept_anonymous:
        description: |-
          AcceptAnonymous forms store submissions without the submitter's
          identity; defaults to true and must be true for public forms.
        example: true
        type: boolean
      content:
        description: Content is the form definition (fields, layout); any JSON value
          but null.
        type: object
      description:
        example: Get in touch with our team
        type: string
      end_date:
        example: "2026-12-31T23:59:59Z"
        type: string
      is_active:
        description: IsActive defaults to true when omitted.
        example: true
        type: boolean
      public_available:
        description: |-
          PublicAvailable forms can be read and filled in without signing in;
          defaults to true.
        example: true
        type: boolean
      slug:
        example: contact-us
        type: string
      start_date:
        example: "2026-01-01T00:00:00Z"
        type: string
      title:
        example: Contact us
        type: string
    type: object
  handlers.FormResponse:
    properties:
      accept_anonymous:
        example: true
        type: boolean
      content:
        type: object
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      description:
        example: Get in touch with our team
        type: string
      end_date:
        example: "2026-12-31T23:59:59Z"
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      public_available:
        example: true
        type: boolean
      slug:
        example: contact-us
        type: string
      start_date:
        example: "2026-01-01T00:00:00Z"
        type: string
      tenant_id:
        example: 1
        type: integer
      title:
        example: Contact us
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    type: object
  handlers.FormSummaryResponse:
    properties:
      accept_anonymous:
        example: true
        type: boolean
      content:
        type: object
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      description:
        example: Get in touch with our team
        type: string
      end_date:
        example: "2026-12-31T23:59:59Z"
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      public_available:
        example: true
        type: boolean
      slug:
        example: contact-us
        type: string
      start_date:
        example: "2026-01-01T00:00:00Z"
        type: string
      submission_count:
        example: 42
        type: integer
      tenant_id:
        example: 1
        type: integer
      title:
        example: Contact us
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
    type: object
...
  handlers.PublicFormResponse:
    properties:
      accept_anonymous:
        description: AcceptAnonymous is false for forms recording who submitted them.
        example: true
        type: boolean
      content:
        type: object
      description:
        example: Get in touch with our team
        type: string
      end_date:
        example: "2026-12-31T23:59:59Z"
        type: string
      public_available:
        description: PublicAvailable is false for forms that require signing in.
        example: true
        type: boolean
      slug:
        example: contact-us
        type: string
      title:
        example: Contact us
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'Username and password of a user: "Basic " + btoa(username + ":" +
    password).'
  title: App API
  version: "1.0"
paths:
...
  /forms:
    get:
      description: Lists the authenticated tenant's forms, newest first, with their
        submission counts.
      parameters:
      - description: Only forms in this state
        in: query
        name: is_active
        type: boolean
      - description: Case-insensitive title substring
        in: query
        name: search
        type: string
      - default: 0
        description: Number of forms to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      - default: 20
        description: Maximum number of forms to return
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Page of forms
          schema:
            $ref: '#/definitions/handlers.FormListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "401":
          description: Missing or invalid API key
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List forms
      tags:
      - forms
    post:
      consumes:
      - application/json
      description: Creates a form owned by the authenticated tenant. The slug must
        be unique across all forms. Forms are public and accept anonymous submissions
        unless public_available or accept_anonymous is false; a public form must accept
        anonymous submissions.
      parameters:
      - description: Form to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.FormRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Form created
          schema:
            $ref: '#/definitions/handlers.FormResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "401":
          description: Missing or invalid API key
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "409":
          description: Slug already exists
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a form
      tags:
      - forms
  /forms/{id}:
    delete:
      description: Deletes one of the authenticated tenant's forms. Forms that have
        submissions or webhooks cannot be deleted.
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Form deleted
        "400":
          description: Invalid form ID
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "401":
          description: Missing or invalid API key
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "404":
          description: Form not found
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "409":
          description: Form has submissions or webhooks
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a form
      tags:
      - forms
    get:
      description: Returns one of the authenticated tenant's forms.
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Form
          schema:
            $ref: '#/definitions/handlers.FormDetailsResponse'
        "400":
          description: Invalid form ID
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "401":
          description: Missing or invalid API key
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "404":
          description: Form not found
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get a form
      tags:
      - forms
    put:
      consumes:
      - application/json
      description: Replaces every field of one of the authenticated tenant's forms.
        Omitted optional fields are cleared; an omitted is_active, public_available
        or accept_anonymous is true.
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: integer
      - description: New state of the form
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.FormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Form updated
          schema:
            $ref: '#/definitions/handlers.FormResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "401":
          description: Missing or invalid API key
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "404":
          description: Form not found
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "409":
          description: Slug already exists
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update a form
      tags:
      - forms
...
  /public/forms/{slug}:
    get:
      description: Returns the form with the given slug if it is active, its tenant
        is active and now is inside its availability window. Forms that are not public_available
        require a signed-in user (bearer token or Basic credentials); credentials
        are optional otherwise.
      parameters:
      - description: Form slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Form
          schema:
            $ref: '#/definitions/handlers.PublicFormResponse'
        "401":
          description: Private form and not signed in, or invalid credentials
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "404":
          description: Form not found or not available
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/errors.HTTPErrorResponse'
      security:
      - BearerAuth: []
      - BasicAuth: []
      summary: Get a form to fill in
      tags:
      - public
...
securityDefinitions:
  ApiKeyAuth:
    description: API key of the tenant, returned when the tenant is created.
    in: header
    name: X-API-Key
    type: apiKey
  BasicAuth:
    type: basic
  BearerAuth:
    description: JWT returned by POST /auth/signin, sent as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"